oops

+++ src/main/java/cn/nukkit/network/protocol/AddHangingEntityPacket.java
@@ -5,7 +5,7 @@ public class AddHangingEntityPacket extends DataPacket {
     public static final byte NETWORK_ID = ProtocolInfo.ADD_HANGING_ENTITY_PACKET;
 
     public long entityUniqueId;
-    public int entityRuntimeId;
+    public long entityRuntimeId;
     public int x;
     public int y;
     public int z;
